+++ veecle-os-runtime/src/datastore/mpsc/mod.rs
@@ -1,0 +1,9 @@
+//! MPSC (multiple-producer, single-consumer) slot implementation.
+
+mod reader;
+mod slot;
+mod writer;
+
+pub use self::reader::Reader;
+pub(crate) use self::slot::Slot;
+pub use self::writer::Writer;